pyproject.toml
[tool.poetry]
name = "repl_python3_Cyber-Junk-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "^3.10.0"
"discord.py" = "^1.7.3"
search-that-hash = "^0.2.8"
replit = "^3.2.4"
bs4 = "^0.0.1"
Flask = "^2.0.1"
google = "^3.0.0"
PyNaCl = "^1.4.0"
pymongo = "^3.12.0"
dnspython = "^2.1.0"
waitress = "^2.0.0"
Quart = "^0.15.1"
uvicorn = "^0.15.0"
fastapi = "^0.68.1"
"quickchart.io" = "^1.0.0"
imageio = "^2.9.0"
Discord-Anti-Spam = "^1.0.2"
python-Levenshtein = "^0.12.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
